telegram-bot/requirements.txt
# Telegram Bot dependencies
python-telegram-bot>=22.5
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0

